ref: refactoring styles CurcularPB
