+++ src/shared/utils/test-users.ts
@@ -1,3 +1,15 @@
+export interface EmployeeGoalPlan {
+	id: number;
+	name: string;
+	position: string;
+	goal: string;
+	date: string;
+	progress: number;
+	taskAll: number;
+	taskDone: number;
+	status: string;
+}
+
 export const testData = [
 	{
 		id: 1,
@@ -6,6 +18,8 @@ export const testData = [
 		goal: 'Карьерный рост',
 		date: '10.01.2024',
 		progress: 25,
+		taskAll: 10,
+		taskDone: 4,
 		status: 'не выполнен',
 	},
 	{
@@ -15,6 +29,8 @@ export const testData = [
 		goal: 'Повышение грейда',
 		date: '01.04.2024',
 		progress: 50,
+		taskAll: 2,
+		taskDone: 1,
 		status: 'в работе',
 	},
 	{
@@ -23,7 +39,9 @@ export const testData = [
 		position: 'Руководитель дирекции',
 		goal: 'Соответствие занимаемой должности',
 		date: '01.07.2024',
-		progress: 75,
+		progress: 80,
+		taskAll: 5,
+		taskDone: 4,
 		status: 'в работе',
 	},
 	{
@@ -33,6 +51,8 @@ export const testData = [
 		goal: 'Развитие софт-скиллов',
 		date: '01.03.2024',
 		progress: 100,
+		taskAll: 3,
+		taskDone: 3,
 		status: 'выполнен',
 	},
 	{
@@ -42,6 +62,8 @@ export const testData = [
 		goal: 'Развитие хард-скиллов',
 		date: '05.02.2024',
 		progress: 100,
+		taskAll: 5,
+		taskDone: 5,
 		status: 'выполнен',
 	},
 	{
@@ -51,6 +73,8 @@ export const testData = [
 		goal: 'Карьерный рост',
 		date: '01.12.2023',
 		progress: 0,
+		taskAll: 3,
+		taskDone: 0,
 		status: 'отменен',
 	},
 	{
@@ -60,6 +84,8 @@ export const testData = [
 		goal: 'Смена специализации',
 		date: '01.08.2024',
 		progress: 0,
+		taskAll: 4,
+		taskDone: 0,
 		status: 'черновик',
 	},
 	{
@@ -69,6 +95,8 @@ export const testData = [
 		goal: 'Получение нового опыта',
 		date: '',
 		progress: 0,
+		taskAll: 6,
+		taskDone: 0,
 		status: 'черновик',
 	},
 	{
@@ -78,6 +106,8 @@ export const testData = [
 		goal: '',
 		date: '',
 		progress: 0,
+		taskAll: 0,
+		taskDone: 0,
 		status: 'отсутствует',
 	},
 	{
@@ -87,6 +117,8 @@ export const testData = [
 		goal: '',
 		date: '',
 		progress: 0,
+		taskAll: 0,
+		taskDone: 0,
 		status: 'отсутствует',
 	},
 	{
@@ -96,6 +128,8 @@ export const testData = [
 		goal: 'Карьерный рост',
 		date: '10.01.2024',
 		progress: 25,
+		taskAll: 4,
+		taskDone: 1,
 		status: 'не выполнен',
 	},
 	{
@@ -105,6 +139,8 @@ export const testData = [
 		goal: 'Повышение грейда',
 		date: '01.04.2024',
 		progress: 50,
+		taskAll: 4,
+		taskDone: 2,
 		status: 'в работе',
 	},
 	{
@@ -113,7 +149,9 @@ export const testData = [
 		position: 'Руководитель дирекции',
 		goal: 'Соответствие занимаемой должности',
 		date: '01.07.2024',
-		progress: 75,
+		progress: 70,
+		taskAll: 10,
+		taskDone: 7,
 		status: 'в работе',
 	},
 	{
@@ -123,6 +161,8 @@ export const testData = [
 		goal: 'Развитие софт-скиллов',
 		date: '01.03.2024',
 		progress: 100,
+		taskAll: 7,
+		taskDone: 7,
 		status: 'выполнен',
 	},
 	{
@@ -132,6 +172,8 @@ export const testData = [
 		goal: 'Развитие хард-скиллов',
 		date: '05.02.2024',
 		progress: 100,
+		taskAll: 8,
+		taskDone: 8,
 		status: 'выполнен',
 	},
 	{
@@ -141,6 +183,8 @@ export const testData = [
 		goal: 'Карьерный рост',
 		date: '01.12.2023',
 		progress: 0,
+		taskAll: 5,
+		taskDone: 0,
 		status: 'отменен',
 	},
 	{
@@ -150,6 +194,8 @@ export const testData = [
 		goal: 'Смена специализации',
 		date: '01.08.2024',
 		progress: 0,
+		taskAll: 4,
+		taskDone: 0,
 		status: 'черновик',
 	},
 	{
@@ -159,6 +205,8 @@ export const testData = [
 		goal: 'Получение нового опыта',
 		date: '',
 		progress: 0,
+		taskAll: 2,
+		taskDone: 0,
 		status: 'черновик',
 	},
 	{
@@ -168,6 +216,8 @@ export const testData = [
 		goal: '',
 		date: '',
 		progress: 0,
+		taskAll: 0,
+		taskDone: 0,
 		status: 'отсутствует',
 	},
 	{
@@ -177,6 +227,8 @@ export const testData = [
 		goal: '',
 		date: '',
 		progress: 0,
+		taskAll: 0,
+		taskDone: 0,
 		status: 'отсутствует',
 	},
 	{
@@ -185,7 +237,9 @@ export const testData = [
 		position: 'Руководитель дирекции',
 		goal: '',
 		date: '',
-		progress: 60,
+		progress: 0,
+		taskAll: 0,
+		taskDone: 0,
 		status: 'отсутствует',
 	},
 ];
